README.md
It is a tool for creating captions for videos is the S2VT model. It creates captions by extracting features from video frames using a data pre-processor script. An attention layer is included in the model design to help with input element focus, along with encoder and decoder layers. To reduce the difference between generated and actual captions, the training script makes advantage of gradient descent and backpropagation. For test video inputs, the testing script creates captions and applies the BLEU score to analyze them. The testing script may be easily executed with the help of a shell script named Hw2_seq2seq.sh. An attention layer is another feature of the model's architecture that allows it to focus on a particular input component.